Part4

+++ part4/bloglist/backend/controllers/bloglist.js
@@ -1,31 +1,55 @@
+/* eslint-disable no-undef */
 const bloglistRouter = require("express").Router()
 const Blog = require("../models/blog")
 
 bloglistRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate("user", { username: 1, name: 1, id: 1 })
   response.json(blogs)
 })
 
 bloglistRouter.post("/", async (request, response) => {
-  const newBlog = new Blog(request.body)
-  
+  const body = request.body
+  const user = request.user
+
+  const newBlog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+  })
+
   if ( !newBlog.title && !newBlog.url ) {
     response
       .status(400)
       .end()
     return
   }
+
   if ( !newBlog.likes ) newBlog.likes = 0
 
-  const newBlogInDb = await newBlog.save()
+  const savedBlog = await newBlog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
 
   response
     .status(201)
-    .json(newBlogInDb)
+    .json(savedBlog)
 })
 
 bloglistRouter.delete("/:id", async (request, response) => {
   const id = request.params.id
+  const blog = await Blog.findById(id)
+
+  const user = request.user
+
+  if ( !(blog.user.toString() === user.id) ) {
+    return response
+      .status(401)
+      .json({ error: "User is not creator of blog or token missing" })
+  }
 
   await Blog.findByIdAndRemove(id)
   response
